Add support for spreadMethod attribute on gradients
Details of changes
L0294-L0297 : Add a simple generic function to transform a bounding box with a matrix.
L1438-L1463 : New variables & reordering of some operations to make possible the new support for spreadMethod.
L1464-L1486 : Some maths to find how many times the gradient must be repeated in the bounding box. The equations where difficult to compute but it appears to work. According to the results of tests, the bounding box of the filled object is transformed with the gradient's inverted matrix. TODO: Because the stroke width does not count in the bounding box, I should maybe add some extra repetition.
L1487-L1491 : Make the gradient longer so that it can be filled with the repetitive gradient stops.
L1492-L1517 : Add a loop and adjust the offsets of the stops as needed. The variable inOrder is here for spreadMethod reflect.