Add support for spreadMethod and href attribute on gradients (#41)
* Add support for spreadMethod, href attributes on gradients:
L0294-L0297 : Add a simple generic function to transform a bounding box with a matrix.
L1451-L1468 : Add support for referencing gradients (`href` attribute).
L1478-L1503 : New variables & reordering of some operations to make possible the new support for `spreadMethod`.
L1504-L1526 : Some maths to find how many times the gradient must be repeated in the bounding box. The equations where difficult to compute but it appears to work. According to the results of tests, the bounding box of the filled object is transformed with the gradient's inverted matrix. TODO: Because the stroke width does not count in the bounding box, I should maybe add some extra repetition.
L1527-L1531 : Make the gradient longer so that it can be filled with the repetitive gradient stops.
L1532-L1557 : Add a loop and adjust the offsets of the stops as needed. The variable `inOrder` is here for `spreadMethod="reflect"`.

* Add support for href attribute on patterns and fix for a test:
L1382-L1399 : Add support for referencing patterns (`href` attribute).
L1401-L1421 : Move most of the calculations inside the `getPattern` function to do them only if necessary. Compute `aspectRatio` after adjusting `width` & `height` => fix test `pservers-pattern-08-f.svg`.

* Upgrade to v0.1.2 and update test results