Correct marker orientation
* SVGtoPDF > BezierSegment > getPointAtLength
This is the buggy function that was causing the issue 43. For bezier curves where the control point and the end point are identical, the slope at that point can't be computed with the derivatives because dx/dt = dy/dt = 0. In this specific case, the vector between the control points can be used.
* SVGtoPDF > SvgElemBasicShape > drawInDocument
For performance reasons, the positions of the markers should only be computed when there are markers to draw